Extract synthetic tests processing
